🧪 Test: add priority test and extract into multiple mapper tests
